% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{calculate_error}
\alias{calculate_error}
\title{Calculate error}
\usage{
calculate_error(data, current_f, current_s, current_g, n_v)
}
\arguments{
\item{data}{list of matrices, data to be factorised,}

\item{current_f}{list of matrices, current F matrices}

\item{current_s}{list of matrices, current S matrices}

\item{current_g}{list of matrices, current G matrices}

\item{n_v}{integer, number of views}
}
\value{
numeric vector, error for each view
}
\description{
Calculate error for ResNMTF
}
