% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{check_boolean}
\alias{check_boolean}
\title{Check boolean inputs}
\usage{
check_boolean(no_clusts, stability, remove_unstable)
}
\arguments{
\item{no_clusts}{boolean, whether to return only the factorisation}

\item{stability}{boolean, whether to perform stability analysis or not}

\item{remove_unstable}{boolean, whether to remove unstable clusters or not}
}
\description{
Check if inputs are boolean
}
