% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_analysis.r
\name{check_empty}
\alias{check_empty}
\title{check for empty matrix}
\usage{
check_empty(data, i)
}
\arguments{
\item{data}{list of matrices to be checked}

\item{i}{index of the view to be checked}
}
\value{
TRUE if any rows/columns with only zeros, FALSE otherwise
}
\description{
check if matrix has any rows/columns with only zeros
}
