% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{check_inputs}
\alias{check_inputs}
\title{Check inputs}
\usage{
check_inputs(
  data,
  init_f,
  init_s,
  init_g,
  k_vec,
  phi,
  xi,
  psi,
  n_iters,
  k_min,
  k_max,
  distance,
  num_repeats,
  no_clusts,
  sample_rate,
  n_stability,
  stability,
  stab_thres,
  remove_unstable
)
}
\arguments{
\item{data}{list of matrices, data to be factorised}

\item{init_f}{list of matrices, initialisation for F matrices}

\item{init_s}{list of matrices, initialisation for S matrices}

\item{init_g}{list of matrices, initialisation for G matrices}

\item{k_vec}{vector of integers, number of clusters to consider}

\item{phi}{list of matrices, default is NULL, restriction matrices for F}

\item{xi}{list of matrices, default is NULL, restriction matrices for S}

\item{psi}{list of matrices, default is NULL, restriction matrices for G}

\item{n_iters}{integer, default is NULL, number of iterations to run for}

\item{k_min}{integer, smallest value of k to be considered initially}

\item{k_max}{integer, largest value of k to be considered initially}

\item{distance}{string, default is "euclidean",
distance metric to use within the bisilhouette score}

\item{num_repeats}{integer, default is 5, minimum value of 2,
number of num_repeats to use for spurious bicluster removal}

\item{no_clusts}{boolean, default is FALSE, whether to return
only the factorisation or not}

\item{sample_rate}{numeric, default is 0.9,
proportion of data to sample for stability analysis}

\item{n_stability}{integer, default is 5,
number of times to repeat stability analysis}

\item{stability}{boolean, default is TRUE,
whether to perform stability analysis or not}

\item{stab_thres}{numeric, default is 0.4,
threshold for stability analysis}

\item{remove_unstable}{boolean, default is TRUE,
whether to remove unstable clusters or not}
}
\value{
list of matrices, data to be factorised
}
\description{
Check if inputs are valid,
normalises and forces data to be non-negative
}
