% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{check_integers}
\alias{check_integers}
\title{Check integers}
\usage{
check_integers(n_iters, k_min, k_max, num_repeats, n_stability)
}
\arguments{
\item{n_iters}{integer, number of iterations to run for}

\item{k_min}{integer, smallest value of k to be considered initially}

\item{k_max}{integer, largest value of k to be considered initially}

\item{num_repeats}{integer, number of num_repeats to use
for spurious bicluster removal}

\item{n_stability}{integer, number of times to repeat stability analysis}
}
\description{
Check if inputs are positive integer
}
