% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{check_lists}
\alias{check_lists}
\title{Check if inputs are list of matrices}
\usage{
check_lists(data, init_f, init_s, init_g)
}
\arguments{
\item{data}{list of matrices, data to be factorised}

\item{init_f}{list of matrices, initialisation for F matrices}

\item{init_s}{list of matrices, initialisation for S matrices}

\item{init_g}{list of matrices, initialisation for G matrices}
}
\value{
list of matrices, data to be factorised
}
\description{
Check if inputs are list of matrices
}
