% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{check_numeric}
\alias{check_numeric}
\title{Check numeric inputs}
\usage{
check_numeric(sample_rate, stab_thres)
}
\arguments{
\item{sample_rate}{numeric, proportion of data to sample for
stability analysis}

\item{stab_thres}{numeric, threshold for stability analysis}
}
\description{
Check if inputs are numeric
}
