% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{check_restriction_mat}
\alias{check_restriction_mat}
\title{Check restriction matrix}
\usage{
check_restriction_mat(matrix, name)
}
\arguments{
\item{matrix}{matrix, restriction matrix}

\item{name}{string, name of the matrix}
}
\description{
Check restriction matrix
}
