% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{extract_bisils}
\alias{extract_bisils}
\title{Extract bisilhouette scores}
\usage{
extract_bisils(res_list, k_vec)
}
\arguments{
\item{res_list}{list of results from ResNMTF}

\item{k_vec}{vector of integers}
}
\value{
vector of bisilhouette scores
}
\description{
Extract bisilhouette scores from results list
}
