% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtain_bicl.r
\name{get_thresholds}
\alias{get_thresholds}
\title{Get thresholds}
\usage{
get_thresholds(x, output_f, repeats, n_views, n_clusts)
}
\arguments{
\item{x}{list of x matrices}

\item{output_f}{list of F matrices}

\item{repeats}{number of repeats}

\item{n_views}{number of views}

\item{n_clusts}{number of biclusters}
}
\value{
list containing average and max scores, as well as the shuffled x
}
\description{
obtain average and max JSD scores between
shuffled F matrices for shuffled data
}
