% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_steps.r
\name{init_mats}
\alias{init_mats}
\title{initialise matrix factors}
\usage{
init_mats(x, k_vec, sigma = 0.05)
}
\arguments{
\item{x}{list of input data}

\item{k_vec}{vector of cluster dimension}

\item{sigma}{noise parameter, default is 0.05}
}
\value{
a list of initial matrices, including
inif_f, init_g, init_s, init_lambda, init_mu
X: list of input data
}
\description{
initialise matrix factors F, S, G etc.
}
