% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_steps.r
\name{init_rest_mats}
\alias{init_rest_mats}
\title{initialise restriction matrices}
\usage{
init_rest_mats(mat, n_v)
}
\arguments{
\item{mat}{matrix to be initialised}

\item{n_v}{number of views}
}
\value{
restriction matrix, shape (n_v, n_v)
}
\description{
if no restrictions required a matrix of zeros is returned,
otherwise the restriction matrix is returned
}
