% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_analysis.r
\name{initial_shuffle}
\alias{initial_shuffle}
\title{shuffle view}
\usage{
initial_shuffle(
  data,
  new_data,
  dims,
  dim,
  i,
  sample_rate,
  row_samples,
  col_samples
)
}
\arguments{
\item{data}{list of matrices to be shuffled}

\item{new_data}{list of shuffled matrices}

\item{dims}{dimensions of the view to be shuffled}

\item{dim}{dimensions of the first view}

\item{i}{index of the view to be shuffled}

\item{sample_rate}{rate at which to sample}

\item{row_samples}{list of rows to sample}

\item{col_samples}{list of cols to sample}

\item{dim_1}{dimensions of the first view}
}
\description{
shuffle view
}
