% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_analysis.r
\name{jaccard_main}
\alias{jaccard_main}
\title{calculate Jaccard index between two biclusterings}
\usage{
jaccard_main(row_c, col_c, true_r, true_c, m, n)
}
\arguments{
\item{row_c}{row clusters}

\item{col_c}{column clusters}

\item{true_r}{true row clusters}

\item{true_c}{true column clusters}

\item{m}{number of row clusters}

\item{n}{number of column clusters}
}
\value{
matrix of Jaccard indices
}
\description{
calculate Jaccard index between two biclusterins
}
