% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{normalisation_check}
\alias{normalisation_check}
\title{Normalise matrix}
\usage{
normalisation_check(current_f, current_g, current_s, n_v)
}
\arguments{
\item{current_f}{list of matrices, current F matrices}

\item{current_g}{list of matrices, current G matrices}

\item{current_s}{list of matrices, current S matrices}

\item{n_v}{integer, number of views}
}
\value{
list of matrices, normalised F, G and S matrices
}
\description{
Normalises matrix factors
}
