% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_analysis.r
\name{number_biclusters}
\alias{number_biclusters}
\title{number of biclusters}
\usage{
number_biclusters(results)
}
\arguments{
\item{results}{results of apply_resnmtf/res_nmtf_inner}
}
\value{
number of biclusters
}
\description{
determine number of biclusters
}
