% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtain_bicl.r
\name{obtain_biclusters}
\alias{obtain_biclusters}
\title{Obtain biclusters}
\usage{
obtain_biclusters(
  data,
  output_f,
  output_g,
  output_s,
  repeats,
  distance = "euclidean"
)
}
\arguments{
\item{data}{list of data matrices}

\item{output_f}{list of F matrices}

\item{output_g}{list of G matrices}

\item{output_s}{list of S matrices}

\item{repeats}{number of repeats}

\item{distance}{distance metric to use, default is "euclidean".
Can be "manhattan", "maximum", "canberra", "binary",
"minkowski" or "pearson"}
}
\value{
list containing row and column clusterings,
}
\description{
obtain biclusters from ResNMTF factorisation,
removing those deemed to be spurious
}
