% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{res_nmtf_inner}
\alias{res_nmtf_inner}
\title{ResNMTF inner functiion}
\usage{
res_nmtf_inner(
  data,
  init_f = NULL,
  init_s = NULL,
  init_g = NULL,
  k_vec = NULL,
  phi = NULL,
  xi = NULL,
  psi = NULL,
  n_iters = NULL,
  repeats = 5,
  distance = "euclidean",
  no_clusts = FALSE
)
}
\arguments{
\item{data}{list of matrices, data to be factorised}

\item{init_f}{list of matrices, initialisation for F matrices}

\item{init_s}{list of matrices, initialisation for S matrices}

\item{init_g}{list of matrices, initialisation for G matrices}

\item{k_vec}{integer, vector of integers, number of clusters to consider}

\item{phi}{list of matrices, default is NULL, restriction matrices for F}

\item{xi}{list of matrices, default is NULL, restriction matrices for S}

\item{psi}{list of matrices, default is NULL, restriction matrices for G}

\item{n_iters}{integer, default is NULL, number of iterations to run for}

\item{repeats}{integer, default is 5, minimum value of 2
number of repeats to use for ??}

\item{distance}{string, default is "euclidean",
distance metric to use within the bisilhouette score}

\item{no_clusts}{boolean, default is FALSE,
whether to return the factorisation
rather than the biclustering}
}
\value{
list of results from ResNMTF
}
\description{
Apply ResNMTF to data, without stability analysis and
for a specific number of biclusters
}
