% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_analysis.r
\name{sample_view}
\alias{sample_view}
\title{sample view}
\usage{
sample_view(data, i, new_data, dim_1, row_samples, col_samples, sample_rate)
}
\arguments{
\item{data}{list of matrices to be shuffled}

\item{i}{index of the view to be shuffled}

\item{new_data}{list of shuffled matrices}

\item{dim_1}{dimensions of the first view}

\item{row_samples}{list of rows to sample}

\item{col_samples}{list of cols to sample}

\item{sample_rate}{rate at which to sample}

\item{dims}{dimensions of the view to be shuffled}
}
\value{
list of shuffled matrices (new_data),
rows (row_samples) and columns (col_samples)
}
\description{
sample view
}
