% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_analysis.r
\name{stability_check}
\alias{stability_check}
\title{stability check}
\usage{
stability_check(
  data,
  results,
  k,
  phi,
  xi,
  psi,
  n_iters,
  num_repeats,
  no_clusts,
  distance,
  sample_rate = 0.9,
  n_stability = 5,
  stab_thres = 0.6,
  remove_unstable = TRUE
)
}
\arguments{
\item{data}{list of input matrices}

\item{results}{results of apply_resnmtf/res_nmtf_inner}

\item{k}{number of biclusters found}

\item{phi}{phi restriction parameter}

\item{xi}{xi restriction parameter}

\item{psi}{psi restriction parameter}

\item{n_iters}{number of iterations}

\item{num_repeats}{number of repeats in removing spurious biclusters}

\item{no_clusts}{number of clusters}

\item{distance}{distance metric to be used within res_nmtf_inner}

\item{sample_rate}{rate at which to sample}

\item{n_stability}{number of stability checks to perform}

\item{stab_thres}{threshold for stability}

\item{remove_unstable}{whether to remove unstable biclusters}
}
\value{
results of biclustering with unstable biclusters removed
(if applicable) or a list containing the initial results and
relevance values for each bicluster
}
\description{
perform stability analysis on the biclustering results
}
