% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_analysis.r
\name{stability_repeat}
\alias{stability_repeat}
\title{stability repeat}
\usage{
stability_repeat(
  results,
  data,
  dim_1,
  k,
  phi,
  xi,
  psi,
  n_iters,
  repeats,
  distance,
  n_views,
  sample_rate
)
}
\arguments{
\item{results}{results of apply_resnmtf/res_nmtf_inner}

\item{data}{list of matrices to be shuffled}

\item{dim_1}{dimensions of the first view}

\item{k}{number of biclusters}

\item{phi}{phi restriction parameter}

\item{xi}{xi restriction parameter}

\item{psi}{psi restriction parameter}

\item{n_iters}{number of iterations}

\item{repeats}{number of repeats}

\item{distance}{distance metric to be used within res_nmtf_inner}

\item{n_views}{number of views}

\item{sample_rate}{rate at which to sample}
}
\value{
matrix (i, j) of relevance values
for each bicluster (j) in each view (i)
and whether stability analysis was performed
(stability_performed)
}
\description{
perform stability analysis on the results
}
