% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{star_prod}
\alias{star_prod}
\title{mat_list: list of k matrices
vec: a vector of length k
Output: a matrix of same dimensions as the entries in mat_list}
\usage{
star_prod(vec, mat_list)
}
\description{
mat_list: list of k matrices
vec: a vector of length k
Output: a matrix of same dimensions as the entries in mat_list
}
