% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_analysis.r
\name{test_cond}
\alias{test_cond}
\title{Test whether there are any columns/rows with only zeros}
\usage{
test_cond(data, attempt)
}
\description{
Test whether there are any columns/rows with only zeros
}
