% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_steps.r
\name{update_f}
\alias{update_f}
\title{Update F matrix}
\usage{
update_f(x, input_f, input_s, input_g, lambda_in, phi, v)
}
\arguments{
\item{x}{list of input data}

\item{input_f}{list of F matrices}

\item{input_s}{S^(v) matrix}

\item{input_g}{G^(v) matrix}

\item{lambda_in}{lambda^(v) vector}

\item{phi}{restriction matrix}

\item{v}{index of the view}
}
\value{
updated F matrix
}
\description{
update F matrix based on the input data
and the current matrix factors
}
