% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_steps.r
\name{update_g}
\alias{update_g}
\title{Update G matrix}
\usage{
update_g(x, input_f, input_s, input_g, mu_in, psi, v)
}
\arguments{
\item{x}{list of input data}

\item{input_f}{F^(v) matrix}

\item{input_s}{S^(v) matrix}

\item{input_g}{list of G matrices}

\item{mu_in}{mu^(v) vector}

\item{psi}{restriction matrix}

\item{v}{index of the view}
}
\value{
updated G matrix
}
\description{
update G matrix based on the input data
and the current matrix factors
}
