% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_steps.r
\name{update_lm}
\alias{update_lm}
\title{Update lambda/mu vectors}
\usage{
update_lm(vec, matrix)
}
\arguments{
\item{vec}{vector to be updated}

\item{matrix}{matrix factor to be used for update}
}
\value{
updated vector
}
\description{
update lambda/mu vectors based on the input data
and the current matrix factors
}
