% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_steps.r
\name{update_matrices}
\alias{update_matrices}
\title{Update matrices}
\usage{
update_matrices(
  x,
  input_f,
  input_s,
  input_g,
  lambda,
  mu,
  phi,
  xi,
  psi,
  n_iters
)
}
\arguments{
\item{x}{list of input data}

\item{input_f}{list of F matrices}

\item{input_s}{list of S matrices}

\item{input_g}{list of G matrices}

\item{lambda}{lambda vector}

\item{mu}{mu vector}

\item{phi}{restriction matrix for F}

\item{xi}{restriction matrix for S}

\item{psi}{restriction matrix for G}

\item{n_iters}{number of iterations}
}
\value{
list containing, ouput_f, output_s, output_g, output_lam, output_mu
}
\description{
update F, S, G matrices based on the input data
and the current matrix factors
}
