% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_steps.r
\name{update_s}
\alias{update_s}
\title{Update S matrix}
\usage{
update_s(x, input_f, input_s, input_g, xi, v)
}
\arguments{
\item{x}{list of input data}

\item{input_f}{F^(v) matrix}

\item{input_s}{list of S matrices}

\item{input_g}{G^(v) matrix}

\item{xi}{restriction matrix}

\item{v}{index of the view}
}
\value{
updated S matrix
}
\description{
update S matrix based on the input data
and the current matrix factors
}
