% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtain_bicl.r
\name{check_biclusters}
\alias{check_biclusters}
\title{Check biclusters}
\usage{
check_biclusters(data, output_f, repeats)
}
\arguments{
\item{data}{list of data matrices}

\item{output_f}{list of F matrices}

\item{repeats}{number of repeats}
}
\value{
list containing the JSD scores,
average and max thresholds
}
\description{
Determine scores and thresholds of the factorisation
to be used to remove biclusters
}
