% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_analysis.r
\name{test_cond}
\alias{test_cond}
\title{Test whether there are any columns/rows with only zeros}
\usage{
test_cond(data, attempt)
}
\arguments{
\item{data}{list of matrices to be tested}

\item{attempt}{number of attempts made to sample data}
}
\value{
TRUE if any columns/rows with only zeros, FALSE otherwise
}
\description{
Test whether there are any columns/rows with only zeros
}
